renames deprecated systems; refactors value_mapping_system name and location
