+++ mescal/visualizations/deprecated_folium_modules/kpi_map_visualizer_base.py
