+++ mescal/visualizations/deprecated_folium_modules/model_visualizer_area_borders.py
@@ -4,8 +4,8 @@ import folium
 import pandas as pd
 from shapely import LineString, Point
 
-from mescal.visualizations.folium_modules_deprecated.model_visualizer_base import StyledModelVisualizerBase
-from mescal.visualizations.styling.icon_styling import BasicArrowIconMap
+from mescal.visualizations.deprecated_folium_modules.model_visualizer_base import StyledModelVisualizerBase
+from mescal.visualizations.deprecated_styling.icon_styling import BasicArrowIconMap
 
 
 class AreaBorderModelVisualizer(StyledModelVisualizerBase):
