+++ mescal/visualizations/deprecated_folium_modules/model_visualizer_areas.py
@@ -4,7 +4,7 @@ import folium
 import pandas as pd
 from shapely import Polygon, MultiPolygon, Point
 
-from mescal.visualizations.folium_modules_deprecated.model_visualizer_base import StyledModelVisualizerBase
+from mescal.visualizations.deprecated_folium_modules.model_visualizer_base import StyledModelVisualizerBase
 
 
 class AreaModelVisualizer(StyledModelVisualizerBase):
