+++ mescal/visualizations/deprecated_folium_modules/model_visualizer_base.py
