+++ mescal/visualizations/deprecated_styling/icon_styling.py
