+++ mescal/visualizations/deprecated_styling/segmented_colormap.py
