+++ mescal/visualizations/deprecated_styling/segmented_interpolator_base.py
