+++ mescal/visualizations/deprecated_styling/segmented_line_width_map.py
