+++ mescal/visualizations/deprecated_styling/segmented_opacity_map.py
