--- mescal/visualizations/styling/__init__.py
@@ -1,3 +1,0 @@
-from mescal.visualizations.styling.segmented_colormap import SegmentedColorMap, SegmentedColorMapLegend
-from mescal.visualizations.styling.segmented_line_width_map import SegmentedLineWidthMap, SegmentedLineWidthMapLegend
-from mescal.visualizations.styling.segmented_opacity_map import SegmentedOpacityMap, SegmentedOpacityMapLegend
